pathfinding seems to work
I currently have a set up where I have Node 1, 2, 5, 7, 8, 9 where all of the costs are currently at 1 and the node number = entity number. The edges are set up such that node 1 connects to 2, 2 connects to 5, 5 connects to 7, 7 connects to 8, 8 connects to 9. The goal set for the pathfinding component is 1 going to 9, so in this particular case, it goes from node 1 through 9 in order to reach the goal. I have another edge that connects node 1 to 9 directly with a cost of 2, and the pathfinding goes from 1 to 9 directly instead of 1 to 9, due to the fact that the total cost is lower than the former, so I guess it  works.